// src/pokeapi.types.ts
export type ShallowLocations = {
    count: number;
    next: string | null;
    previous: string | null;
    results: {name: string; url: string;}[];
};